Backend (Python)/requirements.txt
langchain_core
langchain-google-genai
python-dotenv
fastapi
uvicorn
pydantic